// src/store/game/game.store.enums.ts
export enum GameActions {
  CREATE_GAME = 'CREATE_GAME',
  RESET = 'RESET',
  LOAD_GAME = 'LOAD_GAME',
  ADD_PLAYER = 'ADD_PLAYER',
  SET_ME = 'SET_ME',
  UPDATE_PLAYER = 'UPDATE_PLAYER',
  UPDATE_ARROWS = 'UPDATE_ARROWS',
  MOVE = 'MOVE',
  SHOOT = 'SHOOT',
  START_RESOLVING_ACTIONS = 'START_RESOLVING_ACTIONS',
  END_RESOLVING_ACTIONS = 'END_RESOLVING_ACTIONS',
  SPECTATE = 'SPECTATE',
  SET_ERROR = 'SET_ERROR',
  ADD_ACTION = 'ADD_ACTION',
  RESET_ACTIONS = 'RESET_ACTIONS',
  SET_MESSAGE = 'SET_MESSAGE',
  START_ROUND = 'START_ROUND',
  SET_WINNER = 'SET_WINNER',
}

export enum GameMutations {
  SET_GAME = 'SET_GAME',
  SET_PLAYERS = 'SET_PLAYERS',
  SET_ME = 'SET_ME',
  SET_OBJECT_TILES = 'SET_OBJECT_TILES',
  SET_ARROWS = 'SET_ARROWS',
  SET_RESOLVING = 'SET_RESOLVING',
  SET_IS_SPECTATING = 'SET_IS_SPECTATING',
  SET_ERROR = 'SET_ERROR',
  SET_ACTIONS = 'SET_ACTIONS',
  SET_MESSAGE = 'SET_MESSAGE',
  SET_ACTIVE_ROUND = 'SET_ACTIVE_ROUND',
  SET_WINNER = 'SET_WINNER',
}